streamlit/requirements.txt
echo "streamlit
opencv-python
tensorflow
pillow
numpy" > requirements.txt
